requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
certifi==2018.11.29
chardet==3.0.4
Click==7.0
configparser==3.5.0
enum34==1.1.6
Flask==1.0.2
futures==3.2.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
music-recommender==1.0.1
pylint==1.9.4
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
